+++ packages/filesystem/src/browser/breadcrumbs/filepath-breadcrumbs-container.ts
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import { Container, interfaces, injectable, inject } from 'inversify';
+import { Container, interfaces, injectable, inject } from '@theia/core/shared/inversify';
 import { TreeProps, ContextMenuRenderer, TreeNode, OpenerService, NodeProps } from '@theia/core/lib/browser';
 import { createFileTreeContainer, FileTreeWidget } from '../';
 import { FileTreeModel, FileStatNode } from '../file-tree';